requirements.txt
huggingface_hub==0.22.2
python-dotenv==1.0.1
